+++ tests/flat.props.js
@@ -3,8 +3,14 @@ import { expect } from 'chai';
 import $ from './data/_.flat'; // FORMS
 
 describe('Check Flat $A placeholder field prop', () => {
-  it('$A username should be equal to "Username Placeholder"', () =>
+  it('$A username placeholder should be equal to "Username Placeholder"', () =>
     expect($.$A.$('username').placeholder).to.be.equal('Username Placeholder'));
+
+  it('$A username type should be equal to "text"', () =>
+    expect($.$A.$('username').type).to.be.equal('text'));
+
+  it('$A username bindings should be equal to "default"', () =>
+    expect($.$A.$('username').bindings).to.be.equal('default'));
 });
 
 describe('Check Flat $B error prop', () => {